New program: local_resolution_filter. Pretty primitive for now, but should be sufficient for use during refinement. See comments in Image::ApplyLocalResolutionFilter.

+++ src/core/empirical_distribution.cpp
@@ -51,7 +51,7 @@ float EmpiricalDistribution::GetSampleSumOfSquares()
 	return sum_of_squared_samples;
 }
 
-float EmpiricalDistribution::GetNumberOfSamples()
+long EmpiricalDistribution::GetNumberOfSamples()
 {
 	return number_of_samples;
 }
